Tidy up new local graph computation.

+++ dolfin/graph/GraphBuilder.h
@@ -70,7 +70,7 @@ namespace dolfin
 
     friend class MeshPartitioning;
 
-    typedef boost::unordered_map<std::vector<std::size_t>, std::size_t>
+    typedef boost::unordered_map<std::vector<std::size_t>, std::int32_t>
       FacetCellMap;
 
     static void
